Moved dev fixtures to common conftest.py
