+++ api/tests/conftest.py
@@ -24,3 +24,18 @@ def client(app):
 @pytest.fixture()
 def runner(app):
     return app.test_cli_runner()
+
+@pytest.fixture()
+def dev_app(monkeypatch):
+    monkeypatch.setenv("ENV", "dev")
+    app = run()
+    app.config.update({
+        "TESTING": True,
+    })
+
+    yield app
+
+
+@pytest.fixture()
+def dev_client(dev_app):
+    return dev_app.test_client()
