+++ api/tests/test_build_flask_app.py
@@ -1,24 +1,3 @@
-import pytest
-
-from app import run
-
-
-@pytest.fixture()
-def dev_app(monkeypatch):
-    monkeypatch.setenv("ENV", "dev")
-    app = run()
-    app.config.update({
-        "TESTING": True,
-    })
-
-    yield app
-
-
-@pytest.fixture()
-def dev_client(dev_app):
-    return dev_app.test_client()
-
-
 def test_build_flask_app_dev(dev_app, dev_client):
     """
     Given the package name
